let async_delete_contact be autodiscovered in a tasks.py file
move async method to dedicated file in Archive app, although all methods regarding brevo are in itou.utils.brevo.
Add tasks.py in utils stuffs may make it harder to find back. So keep it next to the command which calls it.
