+++ itou/archive/management/commands/notify_archive_users.py
@@ -10,6 +10,7 @@ from sentry_sdk.crons import monitor
 
 from itou.approvals.models import Approval
 from itou.archive.models import AnonymizedApplication, AnonymizedJobSeeker
+from itou.archive.tasks import async_delete_contact
 from itou.companies.enums import CompanyKind
 from itou.companies.models import JobDescription
 from itou.eligibility.models import EligibilityDiagnosis, GEIQEligibilityDiagnosis
@@ -19,7 +20,6 @@ from itou.job_applications.enums import JobApplicationState
 from itou.job_applications.models import JobApplication, JobApplicationTransitionLog
 from itou.users.models import User, UserKind
 from itou.users.notifications import ArchiveUser, InactiveUser
-from itou.utils.brevo import async_delete_contact
 from itou.utils.command import BaseCommand
 from itou.utils.constants import GRACE_PERIOD, INACTIVITY_PERIOD
